Extract circuit-codegen-annotations module to keep CircuitInject separate
