+++ circuit-codegen-annotations/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=circuit-codegen-annotations
+POM_NAME=Circuit (Codegen annotations)
+POM_DESCRIPTION=Circuit (Codegen annotations)
