+++ samples/star/src/main/kotlin/com/slack/circuit/star/home/AboutScreen.kt
@@ -33,9 +33,9 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.slack.circuit.CircuitInject
 import com.slack.circuit.CircuitUiState
 import com.slack.circuit.Screen
+import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.star.R
 import com.slack.circuit.star.di.AppScope
 import kotlinx.parcelize.Parcelize
@@ -45,11 +45,17 @@ object AboutScreen : Screen {
   object State : CircuitUiState
 }
 
-@CircuitInject(screen = AboutScreen::class, scope = AppScope::class)
+@com.slack.circuit.codegen.annotations.CircuitInject(
+  screen = AboutScreen::class,
+  scope = AppScope::class
+)
 @Composable
 fun AboutPresenter(): AboutScreen.State = AboutScreen.State
 
-@CircuitInject(screen = AboutScreen::class, scope = AppScope::class)
+@com.slack.circuit.codegen.annotations.CircuitInject(
+  screen = AboutScreen::class,
+  scope = AppScope::class
+)
 @Composable
 fun About(modifier: Modifier = Modifier) {
   Scaffold(
