+++ samples/star/src/main/kotlin/com/slack/circuit/star/home/HomePresenter.kt
@@ -30,12 +30,12 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import com.slack.circuit.CircuitContent
-import com.slack.circuit.CircuitInject
 import com.slack.circuit.CircuitUiEvent
 import com.slack.circuit.CircuitUiState
 import com.slack.circuit.NavEvent
 import com.slack.circuit.Navigator
 import com.slack.circuit.Screen
+import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.onNavEvent
 import com.slack.circuit.star.di.AppScope
 import com.slack.circuit.star.home.HomeScreen.Event.ChildNav
@@ -56,7 +56,10 @@ object HomeScreen : Screen {
   }
 }
 
-@CircuitInject(screen = HomeScreen::class, scope = AppScope::class)
+@com.slack.circuit.codegen.annotations.CircuitInject(
+  screen = HomeScreen::class,
+  scope = AppScope::class
+)
 @Composable
 fun HomePresenter(navigator: Navigator): HomeScreen.State {
   val homeNavState = HomeNavPresenter()
@@ -68,7 +71,10 @@ fun HomePresenter(navigator: Navigator): HomeScreen.State {
   }
 }
 
-@CircuitInject(screen = HomeScreen::class, scope = AppScope::class)
+@com.slack.circuit.codegen.annotations.CircuitInject(
+  screen = HomeScreen::class,
+  scope = AppScope::class
+)
 @Composable
 fun HomeContent(state: HomeScreen.State, modifier: Modifier = Modifier) {
   val systemUiController = rememberSystemUiController()
