+++ samples/star/src/main/kotlin/com/slack/circuit/star/petdetail/PetPhotoCarousel.kt
@@ -51,10 +51,10 @@ import com.google.accompanist.pager.HorizontalPagerIndicator
 import com.google.accompanist.pager.PagerState
 import com.google.accompanist.pager.calculateCurrentOffsetForPage
 import com.google.accompanist.pager.rememberPagerState
-import com.slack.circuit.CircuitInject
 import com.slack.circuit.CircuitUiState
 import com.slack.circuit.Presenter
 import com.slack.circuit.Screen
+import com.slack.circuit.codegen.annotations.CircuitInject
 import com.slack.circuit.star.di.AppScope
 import com.slack.circuit.star.petdetail.PetPhotoCarouselTestConstants.CAROUSEL_TAG
 import dagger.assisted.Assisted
@@ -107,7 +107,10 @@ constructor(@Assisted private val screen: PetPhotoCarouselScreen) :
 
   @Composable override fun present() = PetPhotoCarouselScreen.State(screen)
 
-  @CircuitInject(PetPhotoCarouselScreen::class, AppScope::class)
+  @com.slack.circuit.codegen.annotations.CircuitInject(
+    PetPhotoCarouselScreen::class,
+    AppScope::class
+  )
   @AssistedFactory
   interface Factory {
     fun create(screen: PetPhotoCarouselScreen): PetPhotoCarouselPresenter
@@ -118,7 +121,7 @@ internal object PetPhotoCarouselTestConstants {
   const val CAROUSEL_TAG = "carousel"
 }
 
-@CircuitInject(PetPhotoCarouselScreen::class, AppScope::class)
+@com.slack.circuit.codegen.annotations.CircuitInject(PetPhotoCarouselScreen::class, AppScope::class)
 @OptIn(ExperimentalPagerApi::class)
 @Composable
 internal fun PetPhotoCarousel(state: PetPhotoCarouselScreen.State) {
